# Re-use simulator memory to avoid allocations.