Recall that a `Simulator` performs a number of simulations. Previously, for every simulation, the `Simulator` created a new `simulator`. This lead to a lot of extra allocations. This PR reduces the number of allocations by re-using a simulator across multiple simulations, re-using memory whenever possible.

The crux of the implementation is the introduction of a `simulator.reset` method that resets a `simulator` for the next simulation.

```
$ go test -run=$^ -bench=. -count=5 | tee /tmp/baseline.txt
$ go test -run=$^ -bench=. -count=5 | tee /tmp/experiment.txt
$ benchstat /tmp/{baseline,experiment}.txt
name                        old time/op    new time/op    delta
NewWorkload-128               1.29µs ± 2%    0.47µs ± 1%   -63.70%  (p=0.008 n=5+5)
NewSimulator-128              24.3µs ± 1%    23.5µs ± 1%    -3.05%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128    37.3µs ± 2%    33.0µs ± 2%   -11.44%  (p=0.008 n=5+5)
Workloads/NoCalls-128         39.6µs ± 4%    36.7µs ± 3%    -7.33%  (p=0.008 n=5+5)
Workloads/OneCall-128         55.8µs ± 3%    50.3µs ± 1%    -9.91%  (p=0.008 n=5+5)

name                        old alloc/op   new alloc/op   delta
NewWorkload-128                 464B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
NewSimulator-128              7.58kB ± 0%    6.14kB ± 0%   -19.01%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128    8.15kB ± 0%    6.66kB ± 0%   -18.35%  (p=0.008 n=5+5)
Workloads/NoCalls-128         8.67kB ± 0%    7.15kB ± 0%   -17.47%  (p=0.008 n=5+5)
Workloads/OneCall-128         10.6kB ± 0%     8.5kB ± 0%   -19.90%  (p=0.008 n=5+5)

name                        old allocs/op  new allocs/op  delta
NewWorkload-128                 3.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
NewSimulator-128                49.0 ± 0%      32.0 ± 0%   -34.69%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128      64.0 ± 0%      47.0 ± 0%   -26.56%  (p=0.008 n=5+5)
Workloads/NoCalls-128           78.0 ± 0%      60.0 ± 0%   -23.08%  (p=0.008 n=5+5)
Workloads/OneCall-128            116 ± 0%        94 ± 0%   -18.97%  (p=0.008 n=5+5)
```